chore: rename `sources/iqfile.rs` to `sources/iq.rs`, add example to readme
