+++ crates/rs162/src/sources/mod.rs
@@ -5,12 +5,12 @@
 //! - NMEA file source for handling NMEA formatted files.
 //! - Timestamped NMEA source, e.g. for live AIS data streams over TCP.
 //!
-pub mod iqfile;
+pub mod iq;
 pub mod nmea;
 pub mod nmea_ts;
 pub mod rtlsdr;
 
-pub use iqfile::{IqFormat, IqSource};
+pub use iq::{IqFormat, IqSource};
 pub use nmea::NmeaFileSource;
 pub use nmea_ts::TimestampedNmeaTcpSource;
 pub use rtlsdr::{RtlSdrConfig, RtlSdrReceiver};
